deps/v8/test/inspector/debugger/embedder-breakpoints-expected.txt

Tests connecting a session with embedder-managed breakpoints.
Running test: testConnectWithEmbedderBreakpoints
Connecting new session with embedder breakpoints...
Enabling runtime and debugger agents...
Script parsed: test.js
Breakpoint resolved: bp_from_embedder at line 2
Evaluating testFunction()...
Execution paused at line 2, hit breakpoints: ["bp_from_embedder"]
Resuming...
Disconnecting session...
